Convert source file to TS and bump to node 16
