+++ action.yml
@@ -1,7 +1,7 @@
 name: 'LFS asset resolver'
 description: 'Takes a set of Git LFS pointer files, resolves their original URL and invokes an processor Lambda on each key to create a set of output images'
 runs:
-  using: 'node12'
+  using: 'node16'
   main: 'dist/index.js'
 inputs:
   muse_product_slug:
