+++ package.json
@@ -17,5 +17,9 @@
     "lodash.zipwith": "^4.2.0",
     "mime-types": "^2.1.27",
     "node-fetch": "^2.6.1"
+  },
+  "devDependencies": {
+    "@types/node": "^18.14.0",
+    "typescript": "^4.9.5"
   }
 }
